chore: rename instructor_views_api to instructor_views
