+++ test_settings.py
@@ -65,7 +65,7 @@ TEMPLATES = [
 PLATFORM_PLUGIN_COMMUNICATIONS_BULK_EMAIL_API_BACKEND = (
     "platform_plugin_communications.edxapp_wrapper.backends.bulk_email_api_p_v1_test"
 )
-PLATFORM_PLUGIN_COMMUNICATIONS_INSTRUCTOR_VIEWS_API_BACKEND = "platform_plugin_communications.edxapp_wrapper.backends.instructor_views_api_p_v1_test"
+PLATFORM_PLUGIN_COMMUNICATIONS_INSTRUCTOR_VIEWS_API_BACKEND = "platform_plugin_communications.edxapp_wrapper.backends.instructor_views_p_v1_test"
 PLATFORM_PLUGIN_COMMUNICATIONS_COURSE_OVERVIEWS_BACKEND = "platform_plugin_communications.edxapp_wrapper.backends.course_overviews_p_v1_test"
 PLATFORM_PLUGIN_COMMUNICATIONS_UTIL_QUERY_BACKEND = (
     "platform_plugin_communications.edxapp_wrapper.backends.util_query_p_v1_test"
